fix warning

+++ src/main/B99D0.c
@@ -11,8 +11,9 @@ void func_800B8F18(void) {
     s32* temp_v1 = (s32*)D_80137E64;
 
     D_80137E64++;
-    temp_v1[0] = 0x06000000;
-    temp_v1[1] = D_800D31B0;
+    // clang-format off
+    temp_v1[0] = 0x06000000; temp_v1[1] = (s32)D_800D31B0;
+    // clang-format on
 }
 
 #pragma GLOBAL_ASM("asm/us/nonmatchings/main/B99D0/func_800B8F48.s")
